% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rp_table.R
\docType{data}
\name{data_rp_table}
\alias{data_rp_table}
\title{Example data for table_rp}
\format{An object of class \code{matrix} with 5 rows and 5 columns.}
\usage{
data_rp_table
}
\description{
An example data source that is used for
demonstrating the rp_table function.
}
\author{
Andrew Ferris
}
\keyword{datasets}
