% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_t_mat_rp.R
\docType{data}
\name{data_t_mat_rp}
\alias{data_t_mat_rp}
\title{Example data for t_mat_rp}
\format{A 3x3 matrix with numeric elements.}
\source{
Simulated
}
\usage{
data(data_t_mat_rp)
}
\description{
An example data source that is used for demonstrating the t_mat_rp function.
}
\author{
Andrew Ferris
}
\keyword{dataset}
