% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_str_dist.R
\name{hier_str_dist}
\alias{hier_str_dist}
\title{Hierarchical String Distance}
\usage{
hier_str_dist(str_a = str_a, str_b = str_b)
}
\arguments{
\item{str_a}{A character vector}

\item{str_b}{A character vector}
}
\value{
distance An integer representing the hierarchical Levenshtein distance between the two strings
}
\description{
This function uses traditional Levenshtein distances in an ordinal framework to formalise an asymmetric distance between String A and String B. It is specific to the work of the VERRC project and therefore caveats that the two strings are of equal length. Therefore insertions and deletions are not applied and it is
}
\author{
Andrew Ferris
}
