% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_table.R
\name{rp_table}
\alias{rp_table}
\title{Create a Sankey Plot from Table}
\arguments{
\item{table}{A table of class matrix with m colums
where the first m - 1 columns are character and the
final m column is numeric/integer.}
}
\value{
A Sankey Plot

The Sankey Plot object, a list of edges and
nodes.
}
\description{
Create a Sankey Plot from Table

This function takes a n x m matrix where the first m - 1
columns are character factors and the final m column is
a numeric value to create a riverplot (sankey plot) over
the riverplot function in the package riverplot. At this
time style information is missing from the riverplot,
with a view to use RColorBrewer to develop palettes
which can be used to make the plotting more clear.
}
\section{Known Problems}{

There is no style input at this time. This means that
all Sankey plots come out as greyscale.
Labels are not wrapped and hard to clear, making the
interpretation of the image difficult.
Variables are also no ordered, meaning that "rivers"
may not be straight even though they should be.
}

\author{
Andrew Ferris
}
