% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mat_rp.R
\name{t_mat_rp}
\alias{t_mat_rp}
\title{Create a Sankey Plot from Transition Matrix}
\arguments{
\item{table}{A table of class matrix with m columns
and n rows where each element is the value of the
movement.}
}
\value{
A Sankey Plot

The Sankey Plot object, a list of edges and
nodes.
}
\description{
Create a Sankey Plot from Transition Matrix
}
\section{Known Problems}{

There is no style input at this time. This means that
all Sankey plots come out as greyscale.
Labels are not wrapped and hard to clear, making the
interpretation of the image difficult.
Variables are also no ordered, meaning that "rivers"
may not be straight even though they should be.
}

\author{
Andrew Ferris
}
